refactor: extract fs helpers into utils folder
