+++ src/fs/read.js
@@ -1,22 +1,6 @@
-import { dirname, resolve } from 'path';
-import { fileURLToPath } from 'url';
-import { access, readFile } from 'fs/promises';
-import { constants } from 'fs';
-
-const getCurrentFileInfo = (fileUrl) => {
-    const __filename = fileURLToPath(fileUrl);
-    const __dirname = dirname(__filename);
-
-    return [__dirname, __filename];
-};
-
-const checkForFileExistence = async (pathToFile) => {
-    try {
-        await access(pathToFile, constants.R_OK);
-    } catch {
-        throw new Error('FS operation failed');
-    }
-};
+import { resolve } from 'path';
+import { readFile } from 'fs/promises';
+import { checkForFileExistence, getCurrentFileInfo } from '../utils/fs.js';
 
 const read = async () => {
     const [dirname] = getCurrentFileInfo(import.meta.url);
